functions creating Get_dat, dim_date, dim_currency, Fact_sales_order are splited into different python files and addedd more tests
